refactor: 체스판 초기화를 BoardMaker에서 한다 - 빈 체스판과 기본적인 형태의 체스판을 생성할 수 있다
